final project feedback
[This comment is a supplement to the feedback in email.]

Generally a pull request is meant to be accepted into a code base. This one should not be. I’m including code changes that are improvements, but also fragments and comments that are meant to be read as commentary, but not added to the repo.

The code comments also include changes to spacing, because this is what my editor does. This is somewhat useful because it demonstrates standard style, but not the main lesson – so look at a couple of these, but then look for lines with more substantive changes.

General code code comments:
- great docs
- great use of doctest and unit testing
- LICENSE – cool!
- using numpy was ambitious, but it was a good choice for this
- There’s several copies of the same functions and very similar-looking code: in functions.py and game.py; and in board.py, game.py, and main.py.
- some similar stretches of code, that could be combined; for example, in the winning and scoring functions. If you’d like to go further, think about how to abbreviate these functions. A couple of tricks: iterate over the different directions (horizontal, vertical, each of the two diagonals); or, write some code that examines the board, and then use the same code to examine the transposed board, so that you only need to check horizontal and one of the two diagonals.